chore: minor cleanup, loosen some aggressive linting settings
