+++ app/.eslintrc
@@ -112,7 +112,6 @@
         // "react-hooks/exhaustive-deps": "warn",
         "semi": [1, "always"],
         "space-before-function-paren": 0,
-        "@typescript-eslint/no-unused-vars": 1,
         "wrap-iife": [2, "inside"]
     },
     "settings": {
@@ -142,7 +141,9 @@
             ],
             "extends": ["prettier", "plugin:@typescript-eslint/recommended"],
             "rules": {
-                "no-var": 0
+                "@typescript-eslint/no-unused-vars": 1,
+                "no-var": 0,
+                "prefer-const": 1
             }
         },
         {
